+++ src/biwheel.py
@@ -12,7 +12,7 @@ from copy import deepcopy
 from datetime import datetime
 import math
 import os
-from show_util import *
+from chart_utils import *
 from utils import open_file
 
 
@@ -276,9 +276,9 @@ class Biwheel:
                 )
                 a1 = planet_data[7] % 90
                 if ang['model'] == 1:
-                    p1 = major_angularity_curve_2(a1)
+                    p1 = major_angularity_curve_midquadrant_background(a1)
                 else:
-                    p1 = major_angularity_curve(a1)
+                    p1 = major_angularity_curve_cadent_background(a1)
                 a2 = abs(chart['cusps'][1] - planet_data[0])
                 if a2 > 180:
                     a2 = 360 - a2
@@ -442,9 +442,9 @@ class Biwheel:
                 )
                 a1 = planet_data[7] % 90
                 if ang['model'] == 1:
-                    p1 = major_angularity_curve_2(a1)
+                    p1 = major_angularity_curve_midquadrant_background(a1)
                 else:
-                    p1 = major_angularity_curve(a1)
+                    p1 = major_angularity_curve_cadent_background(a1)
                 a2 = abs(chart['cusps'][1] - planet_data[0])
                 if a2 > 180:
                     a2 = 360 - a2