+++ src/chart_options.py
@@ -37,7 +37,16 @@ class ChartOptions(Frame):
         self.bgcurve = Radiogroup(self)
         Radiobutton(self, self.bgcurve, 0, 'At Cadent Cusps', 0.2, 0.2, 0.2)
         Radiobutton(self, self.bgcurve, 1, 'At Mid-quadrant', 0.4, 0.2, 0.2)
-        self.nobg = Checkbutton(self, "Don't Mark", 0.6, 0.2, 0.15)
+        Radiobutton(
+            self,
+            self.bgcurve,
+            2,
+            'Single-equation "Eureka" curve',
+            0.6,
+            0.2,
+            0.2,
+        )
+        self.nobg = Checkbutton(self, "Don't Mark", 0.8, 0.2, 0.15)
         Label(self, 'C1 Orb', 0.225, 0.25, 0.1)
         Label(self, 'C2 0rb', 0.325, 0.25, 0.1)
         Label(self, 'C3 Orb', 0.425, 0.25, 0.1)