+++ src/locations.py
@@ -9,7 +9,7 @@
 
 import json
 from init import *
-from show_util import *
+from chart_utils import *
 from widgets import *