+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Refactor uniwheels
 - Add vertical scaling to title screen
 - Add text wrap to buttons
+- Enable single-formula "Eureka" angularity curve
 
 ### 0.4.13
 - Move to semantic versioning (major.minor.patch format)